お気に入り動画の更新処理に「要約・視聴状況・カテゴリ」の更新を追加

+++ server/updatefavoritevideo/controller/UpdateFavoriteVideoController.ts
@@ -2,7 +2,7 @@ import { Router, Request, Response, NextFunction } from 'express';
 import ENV from '../../env.json';
 import { RouteController } from '../../router/controller/RouteController';
 import { AsyncErrorHandler } from '../../router/service/AsyncErrorHandler';
-import { HTTP_STATUS_CREATED, HTTP_STATUS_OK, HTTP_STATUS_UNPROCESSABLE_ENTITY } from '../../util/const/HttpStatusConst';
+import { HTTP_STATUS_CREATED, HTTP_STATUS_NO_CONTENT, HTTP_STATUS_OK, HTTP_STATUS_UNPROCESSABLE_ENTITY } from '../../util/const/HttpStatusConst';
 import { ApiResponse } from '../../util/service/ApiResponse';
 import { ZodIssue } from 'zod';
 import { FrontUserIdModel } from '../../internaldata/common/properties/FrontUserIdModel';
@@ -15,6 +15,7 @@ import { UpdateFavoriteVideoRequestType } from '../model/UpdateFavoriteVideoRequ
 import { UpdateFavoriteVideoRequestModelSchema } from '../model/UpdateFavoriteVideoRequestModelSchema';
 import { UpdateFavoriteVideoRequestModel } from '../model/UpdateFavoriteVideoRequestModel';
 import { VideoIdModel } from '../../internaldata/common/properties/VideoIdModel';
+import { UpdateFavoriteVideoResponseModel } from '../model/UpdateFavoriteVideoResponseModel';
 
 
 export class UpdateFavoriteVideoController extends RouteController {
@@ -73,10 +74,23 @@ export class UpdateFavoriteVideoController extends RouteController {
         // トランザクション開始
         PrismaTransaction.start(async (tx: Prisma.TransactionClient) => {
 
+            // お気に入り動画更新の永続ロジックを取得
+            const getUpdateFavoriteVideoRepository = this.updateFavoriteVideoService.getUpdateFavoriteVideoRepository();
             // お気に入り動画カテゴリの永続ロジックを取得
             const favoriteVideoCategoryRepository = this.updateFavoriteVideoService.getFavoriteVideoCategoryRepository();
             // お気に入り動画の永続ロジックを取得
             const favoriteVideoRepository = this.updateFavoriteVideoService.getFavoriteVideoRepository();
+
+            // 動画の存在チェック
+            const isExistFavoriteVideoMemo = await this.updateFavoriteVideoService.checkExistFavoriteVideo(
+                getUpdateFavoriteVideoRepository,
+                updateFavoriteVideoRequestModel,
+                frontUserIdModel);
+
+            // お気に入り動画が存在しない
+            if (!isExistFavoriteVideoMemo) {
+                return ApiResponse.create(res, HTTP_STATUS_NO_CONTENT, `お気に入り動画情報が存在しません。`)
+            }
 
             // カテゴリを削除
             await this.updateFavoriteVideoService.deleteCategory(
@@ -87,7 +101,7 @@ export class UpdateFavoriteVideoController extends RouteController {
             );
 
             // カテゴリを登録
-            await this.updateFavoriteVideoService.insertCategory(
+            const favoriteCategoryList = await this.updateFavoriteVideoService.insertCategory(
                 favoriteVideoCategoryRepository,
                 updateFavoriteVideoRequestModel,
                 frontUserIdModel,
@@ -95,14 +109,20 @@ export class UpdateFavoriteVideoController extends RouteController {
             );
 
             // お気に入り動画情報を更新
-            await this.updateFavoriteVideoService.updateFavoriteVideo(
+            const favoriteVideo = await this.updateFavoriteVideoService.updateFavoriteVideo(
                 favoriteVideoRepository,
                 updateFavoriteVideoRequestModel,
                 frontUserIdModel,
                 tx
             );
 
-            return ApiResponse.create(res, HTTP_STATUS_OK, `動画情報の更新が完了しました。`);
+            // レスポンス
+            const updateFavoriteVideoResponseModel = new UpdateFavoriteVideoResponseModel(
+                favoriteVideo,
+                favoriteCategoryList,
+            );
+
+            return ApiResponse.create(res, HTTP_STATUS_OK, `動画情報の更新が完了しました。`, updateFavoriteVideoResponseModel.data);
         }, next);
     }
 }
